backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
ultralytics==8.2.80
